CM-48: Concurrent Collections
