+++ collections-csharp/ConcurrentCollectionExample/ConcurrentCollectionExample.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp", "ExampleApp\ExampleApp.csproj", "{C89A43B6-4DED-4A84-A201-11311807B32A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{ACE1AFD4-1381-460A-ACF7-3D1BF7222F08}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C89A43B6-4DED-4A84-A201-11311807B32A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C89A43B6-4DED-4A84-A201-11311807B32A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C89A43B6-4DED-4A84-A201-11311807B32A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C89A43B6-4DED-4A84-A201-11311807B32A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ACE1AFD4-1381-460A-ACF7-3D1BF7222F08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ACE1AFD4-1381-460A-ACF7-3D1BF7222F08}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ACE1AFD4-1381-460A-ACF7-3D1BF7222F08}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ACE1AFD4-1381-460A-ACF7-3D1BF7222F08}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
