+++ collections-csharp/ConcurrentCollectionExample/ExampleApp/ConcurrentBagExamples/StorageConcurrentBag.cs
