+++ collections-csharp/ConcurrentCollectionExample/ExampleApp/ConcurrentStackExamples/PaperConcurrentStack.cs
