+++ collections-csharp/ConcurrentCollectionExample/ExampleApp/DictionaryExamples/RoleManagerConcurrentDictionary.cs
