+++ collections-csharp/ConcurrentCollectionExample/ExampleApp/DictionaryExamples/RoleManagerDictionary.cs
