+++ collections-csharp/ConcurrentCollectionExample/ExampleApp/Helpers/UserProvider.cs
