+++ collections-csharp/ConcurrentCollectionExample/Tests/PrinterColorPaperConcurrentStackTests.cs
