+++ collections-csharp/ConcurrentCollectionExample/Tests/RoleManagerConcurrentDictionaryTests.cs
