+++ collections-csharp/ConcurrentCollectionExample/Tests/RoleManagerDictionaryTests.cs
